requirements.txt
Flask == 3.0.2
pandas == 2.2.1
PyJwt == 2.8.0
pymongo == 4.6.2